+++ src/tools/index.ts
@@ -19,6 +19,7 @@ import { comparePages } from './compare-pages.js';
 import { getFile } from './get-file.js';
 import { getRevision } from './get-revision.js';
 import { getCategoryMembers } from './get-category-members.js';
+import { getLinksHere } from './get-links-here.js';
 import { listWikis } from './list-wikis.js';
 import { extensionPacks } from './extensions/index.js';
 import { createPage } from './create-page.js';
@@ -52,6 +53,7 @@ const standardTools: Tool<any>[] = [
 	getFile,
 	getRevision,
 	getCategoryMembers,
+	getLinksHere,
 	listWikis,
 	createPage,
 	updatePage,